auth-service/requirements.txt
fastapi>=0.100.0
uvicorn
pymongo
passlib[bcrypt]==1.7.4
python-jose
python-multipart
pydantic>=2.0.0
bcrypt==3.2.0